Dispose instances of Nan::Callback in ~StreamingWorker
StreamingWorkerWrapper explicitly creates three instances of the
Nan::Callback: "callback" (which passed to the AsyncWorker constructor),
"progress" and "error_callback". The calling code doesn't manage their
lifetime and StreamingWorker too. Only "callback" pretends to be
disposed in the destructor of the Nan::AsyncWorker, but other instances
are not – it looks like a memory leak: Nan::Callback leaks, and
v8::Persistent owned by Nan::Callback leaks too.

This commit adds the explicit disposal of "progress" and "error_calback"
instances in the destructor of the StreamingWorker, which looks like
great candidate to manage their lifetimes.